[backend/frontend] chore: implement review comments
