+++ apps/portal-api/src/modules/registration/registration.app.test.ts
@@ -345,7 +345,7 @@ describe('Registration app', () => {
         '@timestamp': '2025-02-03T13:12:15.000Z',
         event_type: TelemetryEventType.REGISTER,
         organization_id: TEST_ORGANIZATIONS.FILIGRAN.ID,
-        organization_name: 'Filigran',
+        organization_name: TEST_ORGANIZATIONS.FILIGRAN.NAME,
         source: TelemetrySource.XTMHUB,
         user_id: TEST_ORGANIZATIONS.FILIGRAN.USERS.BYPASS.ID,
         platform_contract: 'EE',
@@ -374,7 +374,7 @@ describe('Registration app', () => {
         '@timestamp': '2025-02-03T13:12:15.000Z',
         event_type: TelemetryEventType.REGISTER,
         organization_id: TEST_ORGANIZATIONS.FILIGRAN.ID,
-        organization_name: 'Filigran',
+        organization_name: TEST_ORGANIZATIONS.FILIGRAN.NAME,
         source: TelemetrySource.XTMHUB,
         user_id: TEST_ORGANIZATIONS.FILIGRAN.USERS.BYPASS.ID,
         platform_contract: 'EE',
