+++ apps/portal-api/src/modules/registration/test/resolver/registration.resolver.isPlatformRegistered.test.ts
@@ -1,0 +1,71 @@
+import { v4 as uuidv4 } from 'uuid';
+import { afterEach, describe, expect, it, vi } from 'vitest';
+import {
+  contextSimple2,
+  INFO,
+  TEST_ORGANIZATIONS,
+} from '../../../../../tests/tests.const';
+import {
+  IsPlatformRegisteredInput,
+  IsPlatformRegisteredResponse,
+  PlatformRegistrationStatus,
+} from '../../../../__generated__/resolvers-types';
+import { UnknownErrorCode } from '../../../../utils/error/error.code';
+import { registrationApp } from '../../registration.app';
+import registrationResolver from '../../registration.resolver';
+
+describe('Query.isPlatformRegistered', () => {
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it('should return the response from registrationApp on success', async () => {
+    // Given
+    const input: IsPlatformRegisteredInput = { platformId: uuidv4() };
+    const appResponse: IsPlatformRegisteredResponse = {
+      status: PlatformRegistrationStatus.Registered,
+      platformTitle: 'My Platform',
+      organization: { id: TEST_ORGANIZATIONS.FILIGRAN.ID },
+    };
+    vi.spyOn(registrationApp, 'isPlatformRegistered').mockResolvedValue(
+      appResponse
+    );
+
+    // When
+    const result = await registrationResolver.Query!.isPlatformRegistered!(
+      {},
+      { input },
+      contextSimple2,
+      INFO
+    );
+
+    // Then
+    expect(registrationApp.isPlatformRegistered).toHaveBeenCalledWith(input);
+    expect(result).toMatchObject({
+      status: PlatformRegistrationStatus.Registered,
+      platformTitle: 'My Platform',
+      organization: { id: TEST_ORGANIZATIONS.FILIGRAN.ID },
+    });
+  });
+
+  it('should throw a mapped GraphQL error with IsPlatformRegisteredUnknownError when the app throws', async () => {
+    // Given
+    const input: IsPlatformRegisteredInput = { platformId: uuidv4() };
+    vi.spyOn(registrationApp, 'isPlatformRegistered').mockRejectedValue(
+      new Error('UNEXPECTED')
+    );
+
+    // When
+    const call = registrationResolver.Query!.isPlatformRegistered!(
+      {},
+      { input },
+      contextSimple2,
+      INFO
+    );
+
+    // Then
+    await expect(call).rejects.toThrow(
+      UnknownErrorCode.IsPlatformRegisteredUnknownError
+    );
+  });
+});
