+++ apps/portal-api/tests/tests.const.ts
@@ -1,3 +1,4 @@
+import { GraphQLResolveInfo } from 'graphql/index';
 import { PortalCapability } from '../src/__generated__/resolvers-types';
 import CapabilityPortal, {
   CapabilityPortalId,
@@ -12,6 +13,8 @@ import {
   CAPABILITY_BYPASS,
   PLATFORM_ORGANIZATION_UUID,
 } from '../src/portal.const';
+
+export const INFO = {} as GraphQLResolveInfo;
 
 export const TEST_ORGANIZATIONS = {
   FILIGRAN: {
